backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
yt-dlp==2025.01.01
langdetect
openai-whisper
python-multipart
requests
sse-starlette
python-dotenv
